Add files via upload
The project aimed to develop a supervised machine-learning classification model that can be used to predict if a tweet is offensive. More specifically, the model will classify the tweet into one of three categories: Not Offensive (NOT), Targeted Insult (TIN), or Untargeted (UNT). To teach the model, a set of 10,592 annotated tweets were provided to use for training and validation. The ultimate goal of the project was to develop a high-performance model through clever feature selection and the choice of optimal parameters. This was done through a systematic process that involved the use of many Python packages, data preprocessing techniques, trial and error, and experimentation.